+++ product_service/src/main/java/com/cargohub/product_service/presentation/dto/request/CreateProductRequestV1.java
@@ -3,6 +3,7 @@ package com.cargohub.product_service.presentation.dto.request;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.PositiveOrZero;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record CreateProductRequestV1(
@@ -18,11 +19,11 @@ public record CreateProductRequestV1(
         Integer stockQuantity,
 
         @PositiveOrZero( message = "가격은 0원 이상이어야 합니다.")
-        Integer price,
+        BigDecimal price,
 
         Boolean sellable
 ) {
-        public CreateProductRequestV1(String name, UUID firmId, UUID hubId, Integer stockQuantity, Integer price) {
+        public CreateProductRequestV1(String name, UUID firmId, UUID hubId, Integer stockQuantity, BigDecimal price) {
                 this(name, firmId, hubId, stockQuantity, price, true);
         }
 }
